requirements.txt
ultralytics==8.0.159
streamlit==1.35.0
opencv-python-headless
Pillow
gdown
